+++ animory/src/main/java/com/daggle/animory/domain/pet/repository/PetRepository.java
@@ -1,30 +1,14 @@
 package com.daggle.animory.domain.pet.repository;
 
 import com.daggle.animory.domain.pet.entity.Pet;
-import com.daggle.animory.domain.pet.entity.PetType;
-import com.daggle.animory.domain.shelter.entity.Province;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
-import org.springframework.data.domain.Slice;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
 
 import java.util.Optional;
 
 public interface PetRepository extends JpaRepository<Pet, Integer> {
-
-    @Query("select p " +
-        "from Pet p " +
-        "join fetch p.shelter s " +
-        "where s.address.province = :province " +
-        "and p.type = :petType")
-    Slice<Pet> findSliceBy(PetType petType, Province province, Pageable searchCondition);
-
-    @Query("select p " +
-        "from Pet p " +
-        "join fetch p.shelter s " +
-        "join fetch p.petVideos pv ")
-    Slice<Pet> findSliceBy(Pageable pageable);
 
     Page<Pet> findPageBy(Pageable pageable);
