+++ animory/src/main/java/com/daggle/animory/domain/shortform/ShortFormService.java
@@ -1,9 +1,9 @@
 package com.daggle.animory.domain.shortform;
 
-import com.daggle.animory.domain.pet.repository.PetRepository;
 import com.daggle.animory.domain.shortform.dto.request.ShortFormSearchCondition;
 import com.daggle.animory.domain.shortform.dto.response.CategoryShortFormPage;
 import com.daggle.animory.domain.shortform.dto.response.HomeShortFormPage;
+import com.daggle.animory.domain.shortform.repository.PetVideoRepository;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.data.domain.Pageable;
@@ -16,13 +16,13 @@ import org.springframework.transaction.annotation.Transactional;
 @RequiredArgsConstructor
 public class ShortFormService {
 
-    private final PetRepository petRepository;
+    private final PetVideoRepository petVideoRepository;
 
     public CategoryShortFormPage getCategoryShortFormPage(final ShortFormSearchCondition searchCondition,
                                                           final Pageable pageable) {
         return CategoryShortFormPage.of(
             buildCategoryPageTitle(searchCondition),
-            petRepository.findSliceBy(
+            petVideoRepository.findSliceBy(
                 searchCondition.type(),
                 searchCondition.area(),
                 pageable
@@ -33,7 +33,7 @@ public class ShortFormService {
     public HomeShortFormPage getHomeShortFormPage(final Pageable pageable) {
         // TODO: 홈 화면 숏폼 영상은 어떤 순서, 어떤 기준으로 보여줄 것인가?
         return HomeShortFormPage.of(
-            petRepository.findSliceBy(pageable) // TODO: 하드코딩된 Page 숫자
+            petVideoRepository.findSliceBy(pageable) // TODO: 하드코딩된 Page 숫자
         );
     }
