+++ animory/src/test/java/com/daggle/animory/domain/pet/PetRepositoryTest.java
@@ -1,15 +1,12 @@
 package com.daggle.animory.domain.pet;
 
 import com.daggle.animory.domain.pet.entity.Pet;
-import com.daggle.animory.domain.pet.entity.PetType;
 import com.daggle.animory.domain.pet.repository.PetRepository;
-import com.daggle.animory.domain.shelter.entity.Province;
 import com.daggle.animory.testutil.datajpatest.DataJpaTestWithDummyData;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.PageRequest;
-import org.springframework.data.domain.Slice;
 import org.springframework.data.domain.Sort;
 
 import java.util.List;
@@ -23,15 +20,6 @@ class PetRepositoryTest extends DataJpaTestWithDummyData {
     private PetRepository petRepository;
 
     @Test
-    void findSliceBy() {
-        final Slice<Pet> slice = petRepository.findSliceBy(PetType.DOG, Province.광주, PageRequest.of(0, 10));
-
-        print(slice.getContent());
-
-        assertThat(slice.getContent()).hasSize(10);
-    }
-
-    @Test
     void findByShelterId() {
         final Integer shelterId = 1;
         final Page<Pet> page = petRepository.findByShelterId(shelterId, PageRequest.of(0, 10));
@@ -40,29 +28,31 @@ class PetRepositoryTest extends DataJpaTestWithDummyData {
         print(pets);
 
         pets.forEach(
-                p -> assertThat(p.getShelter().getId()).isEqualTo(shelterId)
+            p -> assertThat(p.getShelter().getId()).isEqualTo(shelterId)
         );
     }
 
     @Test
     void findProfilesWithProtectionExpirationDate() {
-        final List<Pet> page = petRepository.findProfilesWithProtectionExpirationDate(PageRequest.of(0, 10)).getContent();
+        final List<Pet> page =
+            petRepository.findProfilesWithProtectionExpirationDate(PageRequest.of(0, 10)).getContent();
 
         print(page);
 
         page.forEach(
-                p -> System.out.println(p.getId() + " " + p.getProtectionExpirationDate())
+            p -> System.out.println(p.getId() + " " + p.getProtectionExpirationDate())
         );
     }
 
     @Test
     void findPage() {
-        final List<Pet> page = petRepository.findPageBy(PageRequest.of(0, 10, Sort.Direction.ASC, "protectionExpirationDate")).getContent();
+        final List<Pet> page = petRepository.findPageBy(PageRequest.of(0, 10, Sort.Direction.ASC,
+                                                                       "protectionExpirationDate")).getContent();
 
         print(page);
 
         page.forEach(
-                p -> System.out.println(p.getId() + " " + p.getProtectionExpirationDate())
+            p -> System.out.println(p.getId() + " " + p.getProtectionExpirationDate())
         );
 
     }
